Changed: event time to 8 byte integer

+++ src/resources/setup.sql
@@ -2,6 +2,6 @@ DROP TABLE events CASCADE;
 
 CREATE TABLE events (
   event_name  VARCHAR(256) NOT NULL,  -- TODO: Create all event list for check
-  event_time  BIGINT          NOT NULL,
-  user_id     INT          NOT NULL   -- TODO: User's table and foreign key
+  event_time  BIGINT       NOT NULL,
+  user_id     BIGINT       NOT NULL   -- TODO: User's table and foreign key
 );
